+++ lib/entity/account.dart
@@ -1,4 +1,5 @@
 import "package:flow/data/flow_icon.dart";
+import "package:flow/data/money.dart";
 import "package:flow/entity/_base.dart";
 import "package:flow/entity/transaction.dart";
 import "package:json_annotation/json_annotation.dart";
@@ -55,14 +56,15 @@ class Account implements EntityBase {
   /// TODO should this be cached?
   @Transient()
   @JsonKey(includeFromJson: false, includeToJson: false)
-  double get balance {
-    return transactions
-        .where((element) => element.transactionDate.isPast)
-        .fold<double>(
-          0,
-          (previousValue, element) => previousValue + element.amount,
-        );
-  }
+  Money get balance => Money(
+        transactions
+            .where((element) => element.transactionDate.isPast)
+            .fold<double>(
+              0,
+              (previousValue, element) => previousValue + element.amount,
+            ),
+        currency,
+      );
 
   Account({
     this.id = 0,