+++ lib/routes/preferences/theme_preferences_page.dart
@@ -1,8 +1,12 @@
 import "package:flow/l10n/extensions.dart";
 import "package:flow/prefs.dart";
-import "package:flow/routes/preferences/theme_preferences/theme_entry.dart";
 import "package:flow/theme/color_themes/registry.dart";
+import "package:flow/theme/helpers.dart";
+import "package:flow/widgets/general/list_header.dart";
+import "package:flow/widgets/theme_petal_selector.dart";
 import "package:flutter/material.dart";
+import "package:material_symbols_icons/symbols.dart";
+// import "package:material_symbols_icons/symbols.dart";
 
 class ThemePreferencesPage extends StatefulWidget {
   const ThemePreferencesPage({super.key});
@@ -11,73 +15,99 @@ class ThemePreferencesPage extends StatefulWidget {
   State<ThemePreferencesPage> createState() => _ThemePreferencesPageState();
 }
 
-class _ThemePreferencesPageState extends State<ThemePreferencesPage>
-    with SingleTickerProviderStateMixin {
-  late final TabController _tabController;
-
+class _ThemePreferencesPageState extends State<ThemePreferencesPage> {
   bool busy = false;
+  bool appIconBusy = false;
+  bool dynamicThemeBusy = false;
 
   @override
   void initState() {
     super.initState();
-    _tabController = TabController(
-      length: 2,
-      vsync: this,
-    );
   }
 
   @override
   Widget build(BuildContext context) {
-    final String? preferencesTheme = LocalPreferences().themeName.get();
-    final String currentTheme = validateThemeName(preferencesTheme)
-        ? preferencesTheme!
-        : lightThemes.keys.first;
+    final String currentTheme = LocalPreferences().getCurrentTheme();
+    final bool themeChangesAppIcon =
+        LocalPreferences().themeChangesAppIcon.get();
+    // final bool enableDynamicTheme = LocalPreferences().enableDynamicTheme.get();
 
     return Scaffold(
       appBar: AppBar(
         title: Text("preferences.theme.choose".t(context)),
-        bottom: TabBar(
-          tabs: [
-            Tab(
-              text: "preferences.theme.light".t(context),
-            ),
-            Tab(
-              text: "preferences.theme.dark".t(context),
-            ),
-          ],
-          controller: _tabController,
-        ),
       ),
       body: SafeArea(
-        child: TabBarView(
-          controller: _tabController,
-          children: [
-            ListView(
-              children: lightThemes.entries
-                  .map(
-                    (entry) => ThemeEntry(
-                      entry: entry,
-                      currentTheme: currentTheme,
-                      handleChange: handleChange,
-                    ),
-                  )
-                  .toList(),
-            ),
-            ListView(
-              children: darkThemes.entries
-                  .map(
-                    (entry) => ThemeEntry(
-                      entry: entry,
-                      currentTheme: currentTheme,
-                      handleChange: handleChange,
-                    ),
-                  )
-                  .toList(),
-            ),
-          ],
+        child: SingleChildScrollView(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Center(
+                child: ThemePetalSelector(
+                  updateOnHover: true,
+                ),
+              ),
+              const SizedBox(height: 16.0),
+              CheckboxListTile.adaptive(
+                title: Text("preferences.theme.themeChangesAppIcon".t(context)),
+                value: themeChangesAppIcon,
+                onChanged: changeThemeChangesAppIcon,
+                secondary: Icon(Symbols.photo_prints_rounded),
+                activeColor: context.colorScheme.primary,
+              ),
+              // CheckboxListTile.adaptive(
+              //   title: Text("preferences.theme.enableDynamicTheme".t(context)),
+              //   value: enableDynamicTheme,
+              //   onChanged: changeEnableDynamicTheme,
+              //   secondary: Icon(Symbols.palette),
+              //   activeColor: context.colorScheme.primary,
+              // ),
+              const SizedBox(height: 16.0),
+              ListHeader(
+                "preferences.theme.other".t(context),
+              ),
+              RadioListTile.adaptive(
+                title: Text(palenight.name),
+                value: "palenight",
+                groupValue: currentTheme,
+                onChanged: (value) => handleChange(value),
+                activeColor: context.colorScheme.primary,
+              ),
+            ],
+          ),
         ),
       ),
     );
+  }
+
+  void changeThemeChangesAppIcon(bool? newValue) async {
+    if (newValue == null) return;
+    if (appIconBusy) return;
+
+    try {
+      appIconBusy = true;
+      await LocalPreferences().themeChangesAppIcon.set(newValue);
+      trySetThemeIcon(newValue ? LocalPreferences().getCurrentTheme() : null);
+    } finally {
+      appIconBusy = false;
+      if (mounted) {
+        setState(() {});
+      }
+    }
+  }
+
+  void changeEnableDynamicTheme(bool? newValue) {
+    if (newValue == null) return;
+    if (dynamicThemeBusy) return;
+
+    try {
+      dynamicThemeBusy = true;
+      LocalPreferences().enableDynamicTheme.set(newValue);
+    } finally {
+      dynamicThemeBusy = false;
+      if (mounted) {
+        setState(() {});
+      }
+    }
   }
 
   void handleChange(String? name) async {
@@ -86,6 +116,9 @@ class _ThemePreferencesPageState extends State<ThemePreferencesPage>
 
     try {
       await LocalPreferences().themeName.set(name);
+      if (LocalPreferences().themeChangesAppIcon.get()) {
+        trySetThemeIcon(name);
+      }
     } finally {
       busy = false;