+++ lib/theme/color_themes/registry.dart
@@ -1,10 +1,12 @@
 import "dart:developer";
+import "dart:io";
 
 import "package:flow/theme/color_themes/default_darks.dart";
 import "package:flow/theme/color_themes/default_lights.dart";
 import "package:flow/theme/color_themes/palenight.dart";
 import "package:flow/theme/flow_color_scheme.dart";
 import "package:flutter/material.dart";
+import "package:flutter_dynamic_icon_plus/flutter_dynamic_icon_plus.dart";
 
 export "default_darks.dart";
 export "default_lights.dart";
@@ -46,12 +48,48 @@ final Map<String, FlowColorScheme> darkThemes = {
   "arcLight": arcLight,
   "driedLilac": driedLilac,
   "neonBoneyard": neonBoneyard,
+};
+
+final Map<String, FlowColorScheme> otherThemes = {
   "palenight": palenight,
 };
+
+final Map<String, String> lightDarkThemeMapping = {
+  "shadeOfViolet": "electricLavender",
+  "blissfulBerry": "pinkQuartz",
+  "cherryPlum": "cottonCandy",
+  "crispChristmasCranberries": "piglet",
+  "burntSienna": "simplyDelicious",
+  "soilOfAvagddu": "creamyApricot",
+  "flagGreen": "yellYellow",
+  "tropicana": "fallGreen",
+  "toyCamouflage": "frostedMintHills",
+  "spreadsheetGreen": "coastalTrim",
+  "tokiwaGreen": "seafairGreen",
+  "hydraTurquoise": "crushedIce",
+  "peacockBlue": "iceEffect",
+  "egyptianBlue": "arcLight",
+  "bohemianBlue": "driedLilac",
+  "spaceBattleBlue": "neonBoneyard",
+};
+
+String? reverseThemeMode(String themeName) {
+  if (lightThemes.containsKey(themeName)) {
+    return lightDarkThemeMapping[themeName];
+  } else if (darkThemes.containsKey(themeName)) {
+    return lightDarkThemeMapping.entries
+        .where((entry) => entry.value == themeName)
+        .firstOrNull
+        ?.key;
+  }
+
+  return null;
+}
 
 final Map<String, FlowColorScheme> allThemes = {
   ...lightThemes,
   ...darkThemes,
+  ...otherThemes,
 };
 
 bool validateThemeName(String? themeName) {
@@ -60,15 +98,48 @@ bool validateThemeName(String? themeName) {
   return allThemes.containsKey(themeName);
 }
 
+bool isThemeDark(String? themeName) {
+  if (themeName == null) return false;
+
+  return darkThemes.containsKey(themeName);
+}
+
 ({FlowColorScheme scheme, ThemeMode mode})? getTheme(String? themeName) {
   if (themeName == null) return null;
 
-  final light = lightThemes[themeName];
-  if (light != null) return (scheme: light, mode: ThemeMode.light);
+  final FlowColorScheme? scheme = allThemes[themeName];
 
-  final dark = darkThemes[themeName];
-  if (dark != null) return (scheme: dark, mode: ThemeMode.dark);
+  if (scheme == null) {
+    log("Unknown theme: $themeName");
+    return null;
+  }
 
-  log("Unknown theme: $themeName");
-  return null;
+  final mode = scheme.isDark ? ThemeMode.dark : ThemeMode.light;
+
+  return (scheme: scheme, mode: mode);
+}
+
+void trySetThemeIcon(String? name) async {
+  name ??= "shadeOfViolet";
+
+  if (!Platform.isIOS) return;
+
+  late final String icon;
+
+  if (lightThemes.containsKey(name)) {
+    icon = name;
+  } else if (darkThemes.containsKey(name)) {
+    icon = lightDarkThemeMapping.keys.firstWhere(
+      (key) => lightDarkThemeMapping[key] == name,
+      orElse: () => "shadeOfViolet",
+    );
+  } else {
+    icon = "shadeOfViolet";
+  }
+
+  try {
+    await FlutterDynamicIconPlus.setAlternateIconName(iconName: icon);
+  } catch (e) {
+    log("Failed to set app icon: $e");
+  }
 }